// admin/types.d.ts
export interface ExpertTraderType {
  _id: string;
  firstname: string;
  lastname: string;
  email: string;
  photo: string;
  comment: string;
  profitShare: string; // e.g., "25%"
  winRate: string; // e.g., "95%"
  createdAt: string; // ISO date string
  updatedAt: string; // ISO date string
  __v: number;
}